chor(user): eslint was applied
