+++ packages/user-account/src/useCases/index.js
@@ -1,5 +1,5 @@
-const {createUser} = require('./user.service');
+const { createUser } = require('./user.service');
 
 module.exports = {
-	createUserServices: createUser,
+  createUserServices: createUser,
 };
