+++ packages/user-account/src/useCases/user.service.js
@@ -1,11 +1,11 @@
-const validatorHandler = require("../utils/validator/validatorHandler");
-const { createUserSchema } = require("../utils/validator/schema/user.schema");
 const boom = require('@hapi/boom');
+const validatorHandler = require('../utils/validator/validatorHandler');
+const { createUserSchema } = require('../utils/validator/schema/user.schema');
 
-function createUser (model) {
+function createUser(model) {
   const data = validatorHandler(createUserSchema, model);
-  if (data != model) {
-    return boom.badRequest(data).details
+  if (data !== model) {
+    return boom.badRequest(data).details;
   }
   return data;
 }
